Support multipage hOCR, add word confidence values
This change adds support for multipage files, i.e. hOCR files in
which `div[@class='ocr_page']` returns a sequence. It takes the first
filename as the `fileName` in the `sourceImageInformation` and creates
a single `<Layout>` for all `<Page>`s.

I added a function that extracts the word confidence as integer,
divides it by 100 to get a float between 0 and 1 and adjusted the
template for words to include it in `@WC`.